Move relation cell tests
